Remove Config.makeConfig
This doesn't really make sense as part of the `Config` API.

We can maybe make class `Utils` public, but, I don't know how useful it
is anyways.
